+++ CHANGES.md
@@ -5,6 +5,12 @@ Waiting on https://github.com/standard/eslint-config-standard/pull/160
 as would like to remove deprecated rule https://eslint.org/docs/rules/no-negated-in-lhs
 -->
 
+## ?
+
+- Enhancement [`sauron.js`] (tighten): For `jsdoc/require-jsdoc`, target
+  `ClassDeclaration`, `ClassExpression`, `MethodDefinition` in addition to
+  `FunctionDeclaration`
+
 ## 16.5.0
 
 - Enhancement [`index.js`] (tighten): Add `jsdoc/no-bad-blocks`